packages/api/pyproject.toml
[tool.poetry]
name = "semantic-search"
version = "0.1.0"
description = ""
authors = ["Matthew Rowland <contact@mathhulk.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
black = "^24.10.0"
pinecone = "^5.3.1"
sentence-transformers = "^3.2.0"
python-dotenv = "^1.0.1"
waitress = "^3.0.0"
flask-cors = "^5.0.0"
faiss-cpu = "^1.9.0"
gdown = "^5.2.0"
pandas = "^2.2.3"

[tool.poetry.scripts]
dev = "semantic_search:dev"
prod = "semantic_search:prod"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
